requirements.txt
protobuf
transformers==4.27.1
cpm_kernels
torch>=1.10
gradio
mdtex2html
sentencepiece
accelerate
paddlepaddle
uvicorn
requests
fastapi
